+++ src/main/java/ru/practicum/booking/BookingService.java
@@ -4,13 +4,13 @@ import java.util.List;
 
 public interface BookingService {
 
-    ResponseBookingDto addBooking(BookingDto bookingDto, int bookerId);
+    ResponseBookingDto addBooking(BookingDto bookingDto, Long bookerId);  // Изменено на Long
 
-    ResponseBookingDto patchBooking(int ownerId, int bookingId, boolean isApproved);
+    ResponseBookingDto patchBooking(Long ownerId, Long bookingId, boolean isApproved);  // Изменено на Long
 
-    ResponseBookingDto getBookingById(int requesterId, int bookingId);
+    ResponseBookingDto getBookingById(Long requesterId, Long bookingId);  // Изменено на Long
 
-    List<ResponseBookingDto> getAllUsersBookings(int usersId, BookingState state);
+    List<ResponseBookingDto> getAllUsersBookings(Long usersId, BookingState state);  // Изменено на Long
 
-    List<ResponseBookingDto> getAllItemOwnerBookings(int ownerId, BookingState state);
+    List<ResponseBookingDto> getAllItemOwnerBookings(Long ownerId, BookingState state);  // Изменено на Long
 }
