+++ src/main/java/ru/practicum/item/ItemController.java
@@ -17,25 +17,25 @@ public class ItemController {
     private final ItemService itemService;
 
     @PostMapping
-    public ItemDto addItem(@RequestBody @Valid ItemDto itemDto, @RequestHeader("X-Sharer-User-Id") int ownerId) {
+    public ItemDto addItem(@RequestBody @Valid ItemDto itemDto, @RequestHeader("X-Sharer-User-Id") Long ownerId) {
         log.info("Получен запрос POST /items");
         return itemService.addItem(itemDto, ownerId);
     }
 
     @PatchMapping("/{itemId}")
-    public ItemDto patchItem(@RequestBody ItemDto itemDto, @PathVariable int itemId, @RequestHeader("X-Sharer-User-Id") int ownerId) {
+    public ItemDto patchItem(@RequestBody ItemDto itemDto, @PathVariable Long itemId, @RequestHeader("X-Sharer-User-Id") Long ownerId) {
         log.info("Получен запрос PATCH /items/{itemId}");
         return itemService.updateItem(itemDto, itemId, ownerId);
     }
 
     @GetMapping("/{itemId}")
-    public ItemDtoBookingsAndComments getItemById(@RequestHeader("X-Sharer-User-Id") int ownerId, @PathVariable int itemId) {
+    public ItemDtoBookingsAndComments getItemById(@RequestHeader("X-Sharer-User-Id") Long ownerId, @PathVariable Long itemId) {
         log.info("Получен запрос GET /items/{itemId}");
         return itemService.getItemById(ownerId, itemId);
     }
 
     @GetMapping
-    public List<ItemDtoBookingsAndComments> getUserItems(@RequestHeader("X-Sharer-User-Id") int ownerId) {
+    public List<ItemDtoBookingsAndComments> getUserItems(@RequestHeader("X-Sharer-User-Id") Long ownerId) {
         log.info("Получен запрос GET /items");
         return itemService.getUserItems(ownerId);
     }
@@ -47,7 +47,7 @@ public class ItemController {
     }
 
     @PostMapping("/{itemId}/comment")
-    public CommentDto addComment(@RequestBody @Valid CommentDto commentDto, @PathVariable int itemId, @RequestHeader("X-Sharer-User-Id") int authorId) {
+    public CommentDto addComment(@RequestBody @Valid CommentDto commentDto, @PathVariable Long itemId, @RequestHeader("X-Sharer-User-Id") Long authorId) {
         log.info("Получен запрос POST /items");
         return itemService.addComment(commentDto, itemId, authorId);
     }
